bcsproject/requirements.txt
amqp==5.0.6
asgiref==3.4.1
astroid==2.4.1
billiard==3.6.4.0
boto3==1.17.112
botocore==1.20.112
celery==5.0.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.2.0
colorama==0.4.4
coreschema==0.0.4
dj-database-url==0.5.0
dj-rest-auth==2.1.9
Django==3.2.5
django-autoslug==1.9.8
django-axes==5.20.0
django-bootstrap-form==3.4
django-cities-light==3.8.1
django-cors-headers==3.7.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-db-logger==0.1.10
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-formtools==2.3
django-hashid-field==3.2.1
django-ipware==3.0.2
django-markdown-deux==1.0.5
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
django-settings-json==1.0.2
django-storages==1.11.1
django-stubs==1.8.0
django-stubs-ext==0.2.0
django-template-preview==0.1.5
django-timezone-field==4.2.1
django-tinymce==3.3.0
django-watchman==1.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
flake8==3.8.4
gunicorn==20.1.0
h11==0.12.0
hashids==1.3.1
idna==2.10
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.1.0
lazy-object-proxy==1.4.3
markdown2==2.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
progressbar2==3.53.1
prompt-toolkit==3.0.8
psycopg2==2.9.1
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.3
pylint-celery==0.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-utils==2.5.5
pytz==2020.1
requests==2.24.0
s3transfer==0.4.2
six==1.15.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.23
sqlparse==0.4.0
toml==0.10.2
typing-extensions==3.7.4.3
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.25.11
uvicorn==0.13.3
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.12.1
